+++ src/VectorSpace.js
@@ -1,18 +1,18 @@
-var inherits = require('inherits')
-var itemsPool = require('./itemsPool')
-var matrixMultiplication = require('matrix-multiplication')
-var operators = require('./operators.json')
-var staticProps = require('static-props')
-var TensorSpace = require('./TensorSpace')
-var toData = require('./toData')
+const inherits = require('inherits')
+const itemsPool = require('./itemsPool')
+const matrixMultiplication = require('matrix-multiplication')
+const operators = require('./operators.json')
+const staticProps = require('static-props')
+const TensorSpace = require('./TensorSpace')
+const toData = require('./toData')
 
 /**
  * Space of vectors
  *
  * ```
- * var V = VectorSpace(R)(2)
+ * const V = VectorSpace(R)(2)
  *
- * var v = new V([1, 2])
+ * const v = new V([1, 2])
  * ```
  *
  * @param {Object} Scalar
@@ -21,9 +21,9 @@ var toData = require('./toData')
  */
 
 function VectorSpace (Scalar) {
-  var addition = Scalar.addition
-  var multiplication = Scalar.multiplication
-  var subtraction = Scalar.subtraction
+  const addition = Scalar.addition
+  const multiplication = Scalar.multiplication
+  const subtraction = Scalar.subtraction
 
   /**
    * @param {Number} dimension
@@ -32,9 +32,9 @@ function VectorSpace (Scalar) {
    */
 
   return function (dimension) {
-    var indices = [dimension]
+    const indices = [dimension]
 
-    var AbstractVector = TensorSpace(Scalar)(indices)
+    const AbstractVector = TensorSpace(Scalar)(indices)
 
     /**
      * Computes the cross product of two vectors.
@@ -48,18 +48,18 @@ function VectorSpace (Scalar) {
      */
 
     function crossProduct (vector1, vector2) {
-      var vectorData1 = toData(vector1)
-      var vectorData2 = toData(vector2)
+      const vectorData1 = toData(vector1)
+      const vectorData2 = toData(vector2)
 
-      var ux = vectorData1[0]
-      var uy = vectorData1[1]
-      var uz = vectorData1[2]
+      const ux = vectorData1[0]
+      const uy = vectorData1[1]
+      const uz = vectorData1[2]
 
-      var vx = vectorData2[0]
-      var vy = vectorData2[1]
-      var vz = vectorData2[2]
+      const vx = vectorData2[0]
+      const vy = vectorData2[1]
+      const vz = vectorData2[2]
 
-      var vector = []
+      let vector = []
 
       vector.push(subtraction(multiplication(uy, vz), multiplication(uz, vy)))
       vector.push(subtraction(multiplication(uz, vx), multiplication(ux, vz)))
@@ -76,10 +76,10 @@ function VectorSpace (Scalar) {
      */
 
     function multiplicationByMatrix (leftVector, rightMatrix) {
-      var leftVectorData = toData(leftVector)
-      var rightMatrixData = toData(rightMatrix)
+      const leftVectorData = toData(leftVector)
+      const rightMatrixData = toData(rightMatrix)
 
-      var rowByColumnMultiplication = matrixMultiplication(Scalar)(dimension)
+      const rowByColumnMultiplication = matrixMultiplication(Scalar)(dimension)
 
       return rowByColumnMultiplication(leftVectorData, rightMatrixData)
     }
@@ -97,11 +97,11 @@ function VectorSpace (Scalar) {
      */
 
     function norm (vector) {
-      var data = toData(vector)
+      const data = toData(vector)
 
-      var value = multiplication(data[0], data[0])
+      let value = multiplication(data[0], data[0])
 
-      for (var i = 1; i < dimension; i++) {
+      for (let i = 1; i < dimension; i++) {
         value = addition(value, multiplication(data[i], data[i]))
       }
 
@@ -121,16 +121,16 @@ function VectorSpace (Scalar) {
 
     function scalarProduct (vector1, vector2) {
       // TODO use tensor product and then contraction (trace)
-      var vectorData1 = toData(vector1)
-      var vectorData2 = toData(vector2)
+      const vectorData1 = toData(vector1)
+      const vectorData2 = toData(vector2)
 
       if (vectorData1.length !== vectorData2.length) {
         throw new TypeError('Vectors have not the same dimension')
       }
 
-      var result = multiplication(vectorData1[0], vectorData2[0])
+      let result = multiplication(vectorData1[0], vectorData2[0])
 
-      for (var i = 1; i < dimension; i++) {
+      for (let i = 1; i < dimension; i++) {
         result = addition(result, multiplication(vectorData1[i], vectorData2[i]))
       }
 
@@ -155,18 +155,18 @@ function VectorSpace (Scalar) {
     staticProps(Vector)({ dimension })
 
     Vector.prototype.scalarProduct = function (vector) {
-      var data = this.data
+      const data = this.data
 
-      var result = scalarProduct(data, vector)
+      const result = scalarProduct(data, vector)
 
       return new Scalar(result)
     }
 
     // Cross product is defined only in dimension 3.
     function crossProductMethod (vector) {
-      var data = this.data
+      const data = this.data
 
-      var result = crossProduct(data, vector)
+      const result = crossProduct(data, vector)
 
       return new Vector(result)
     }
@@ -179,18 +179,18 @@ function VectorSpace (Scalar) {
     }
 
     Vector.prototype.multiplication = function (rightMatrix) {
-      var MatrixSpace = itemsPool.get('MatrixSpace')
+      const MatrixSpace = itemsPool.get('MatrixSpace')
 
-      var leftVectorData = this.data
-      var result = multiplicationByMatrix(leftVectorData, rightMatrix)
+      const leftVectorData = this.data
+      const result = multiplicationByMatrix(leftVectorData, rightMatrix)
 
       // TODO rightNumRows equals dimension
       // but the vector should be transposed.
       // Add transpose operator for vectors, then use it implicitly.
-      var rightNumRows = dimension
-      var rightNumCols = result.length / rightNumRows
+      const rightNumRows = dimension
+      const rightNumCols = result.length / rightNumRows
 
-      var Matrix = MatrixSpace(Scalar)(rightNumRows, rightNumCols)
+      const Matrix = MatrixSpace(Scalar)(rightNumRows, rightNumCols)
 
       return new Matrix(result)
     }
@@ -218,7 +218,7 @@ function VectorSpace (Scalar) {
     Vector.mul = multiplicationByMatrix
     Vector.prototype.mul = Vector.prototype.multiplication
 
-    var myOperators = ['scalarProduct'].concat(operators.group)
+    const myOperators = ['scalarProduct'].concat(operators.group)
 
     myOperators.forEach((operator) => {
       operators.aliasesOf[operator].forEach((alias) => {