+++ src/realField.js
@@ -1,4 +1,4 @@
-var realField = {
+const realField = {
   zero: 0,
   one: 1,
   // NaN, Infinity and -Infinity are not allowed.