+++ src/toData.js
@@ -1,4 +1,4 @@
-var no = require('not-defined')
+const no = require('not-defined')
 
 /**
  * Extract data attribute, if any, and check it
@@ -9,7 +9,7 @@ var no = require('not-defined')
  */
 
 function toData (arg) {
-  var data
+  let data
 
   if (no(arg.data)) data = arg
   else data = arg.data